pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-groq>=0.3.4",
    "langchain-mcp-adapters>=0.1.7",
    "langgraph>=0.5.0",
    "mcp>=1.10.1",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "fastapi>=0.112.0",
    "uvicorn>=0.35.0",
    "pydantic>=2.7.0",
]
